refactor(clustering): move grouping algorithms into package
Keep the existing cluster_helpers import surface while making meta-graph,
superclustering, and anchored grouping owned by static_analyzer.clustering.

Amp-Thread-ID: https://ampcode.com/threads/T-01a01478-33f2-70cf-8a11-9475c7d3a4cd
